Committing initial progress
Wanted to get the base for a page ready, so that additions can be more granular. So far, the instructions say how to add the CatenaPartiesManager to the scene, and use it to create a party - with the warning of needing an account first. This starts stubbing out the code needed to join a party.