Adding a page for demo documentation
The goal of this page is not to explain how to use the various services in Unity (which is covered in the other pages), but rather just as a guide for how to test the Catena services using the demo. This is meant to be a companion to the demo so that the functionality can be more easily observed, and provide a good example of how integrating the Unity SDK into a project might look.