Adding documentation for strategies and hooks
The new page covers both strategies and hooks, as the two are similar and related - but makes sure to explain the differences. The goal of the new page is to explain for both strategies and hooks what they are, what our existing hooks and strategies do, and how you would go about modifying or creating your own.

This is first draft - so let me know if there's any grammar or wording errors that I have missed! (Or if I just got anything wrong)